Continue to solve complexity by moving interfaces to types folder, seeking for the reusable code and moving it to components folder. Optimized components, e.g. Search. Creating index.ts file for analytics to have ability export several components together. Also a little bit changes in Analytics page and making components for it, but it haven't been completed yet
