+++ src/components/pages/analytics/index.ts
@@ -1,0 +1,4 @@
+export { default as SaleDetailsModal } from './SaleDetailsModal'
+export { default as SaleItem } from './SaleItem'
+export { default as SalesDynamicsChart } from './SalesDynamicsChart'
+export { default as TopSalesChart } from './TopSalesChart'
