+++ src/components/product/category/ProductCategoriesList.tsx
@@ -1,24 +1,18 @@
-import { ProductTree } from '@/types/product'
+import { ProductCategoriesProps } from '@/types/product'
 import React from 'react'
 import ProductCategoryItem from './ProductCategoryItem'
-
-interface ProductCategoriesProps {
-	currentLevel: any
-	setEditingProduct: (product: any) => void
-	handleItemClick: (key: string) => void
-}
 
 const ProductCategoriesList: React.FC<ProductCategoriesProps> = ({
 	currentLevel,
 	setEditingProduct,
 	handleItemClick,
 }) => {
-    return (
+	return (
 		<>
 			{Object.entries(currentLevel).map(([category, products]) => (
 				<ProductCategoryItem
-                    key={category}
-                    category={category}
+					key={category}
+					category={category}
 					products={products}
 					setEditingProduct={setEditingProduct}
 					handleItemClick={handleItemClick}
