+++ src/pages/index.ts
@@ -1,4 +1,4 @@
-export { default as AnalyticsPage } from './AnalyticsPage'
+export { default as AnalyticsPage } from './analytics-page/AnalyticsPage'
 export { default as CreateProductPage } from './CreateProductPage'
 export { default as HomePage } from './HomePage'
 export { default as ProductArrivalPage } from './ProductArrivalPage'
