+++ src/styles/pages/create-product/CreateProductPage.scss
@@ -1,146 +1,73 @@
 .form-container {
-    max-width: 800px;
-    margin: 0 auto;
-    padding: 20px;
-    background-color: #f9f9f9;
-    border-radius: 8px;
-    box-shadow: 0 4px 10px rgba(0, 0, 0, 0.1);
+	@apply max-w-2xl mx-auto p-4 bg-white rounded-lg shadow-md;
 }
 
 h2 {
-    text-align: center;
-    color: #333;
-    margin-bottom: 20px;
+	@apply text-center text-2xl font-bold mb-5;
 }
 
 .form-group {
-    margin-bottom: 15px;
+	@apply mb-4;
 }
 
 label {
-    font-weight: bold;
-    display: block;
-    margin-bottom: 8px;
-    color: #444;
+	@apply block font-bold mb-2 text-gray-700;
 }
 
-input, select {
-    width: 100%;
-    padding: 8px;
-    font-size: 16px;
-    border: 1px solid #ccc;
-    border-radius: 4px;
-    box-sizing: border-box;
+input,
+select {
+	@apply w-full p-2 border border-gray-300 rounded-md;
 }
 
 select {
-    -webkit-appearance: none;
-    -moz-appearance: none;
-    appearance: none;
+	@apply appearance-none;
 }
 
 button.submit-btn {
-    width: 100%;
-    padding: 12px;
-    background-color: #007bff;
-    color: white;
-    border: none;
-    border-radius: 5px;
-    font-size: 16px;
-    cursor: pointer;
-
-    &:hover {
-        background-color: #0056b3;
-    }
+	@apply w-full p-3 bg-blue-500 text-white rounded-md hover:bg-blue-600;
 }
 
-input[type="number"] {
-    -moz-appearance: textfield;
-
-    &::-webkit-outer-spin-button,
-    &::-webkit-inner-spin-button {
-        -webkit-appearance: none;
-        margin: 0;
-    }
+input[type='number'] {
+	@apply appearance-none;
 }
 
 .switcher {
-    display: flex;
-    justify-content: center;
-    column-gap: 50px;
-    margin: 40px auto;
-    background-color: #3a8ae0;
-    border-radius: 10px;
-    width: fit-content;
-    padding: 20px 20px 30px 20px;
+	@apply flex justify-center gap-10 bg-blue-500 p-4 rounded-lg;
 }
 
 .switcher-value {
-    font-size: 24px;
-    color: #fff;
-    position: relative;
-    cursor: pointer;
-    transition: all .3s;
+	@apply text-2xl text-white cursor-pointer transition duration-300;
+	&:hover {
+		&::after {
+			@apply block absolute mt-2 w-full h-1 bg-white rounded-full;
+		}
+	}
 
-    &:hover {
-        &::after {
-            content: '';
-            position: absolute;
-            display: block;
-            margin-top: 10px;
-            width: 100%;
-            height: 5px;
-            background-color: #fff;
-            border-radius: 30px;
-        }
-    }
-
-    &.active {
-        &::after {
-            content: '';
-            position: absolute;
-            display: block;
-            margin-top: 10px;
-            width: 100%;
-            height: 5px;
-            background-color: #fff;
-            border-radius: 30px;
-        }
-    }
+	&.active {
+		&::after {
+			@apply block absolute mt-2 w-full h-1 bg-white rounded-full;
+		}
+	}
 }
 
 #form-many {
-    .form-title {
-        text-align: center;
-        color: #333;
-        margin-bottom: 20px;
-    }
+	.form-title {
+		@apply text-center text-2xl font-bold mb-5;
+	}
 
-    .form-instructions {
-        font-weight: bold;
-        color: #444;
-        margin-top: 10px;
+	.form-instructions {
+		@apply font-bold text-gray-700 mt-2;
 
-        &-main {
-            display: block;
-            font-size: 18px;
-        }
-    }
+		&-main {
+			@apply block text-xl;
+		}
+	}
 
-    .multiple-products-form {
-        display: flex;
-        flex-direction: column;
-        align-items: center;
+	.multiple-products-form {
+		@apply flex flex-col items-center;
 
-        .products-textarea {
-            width: 100%;
-            height: 150px;
-            padding: 10px;
-            font-size: 16px;
-            border: 1px solid #ccc;
-            border-radius: 4px;
-            box-sizing: border-box;
-            margin-bottom: 20px;
-        }
-    }
+		.products-textarea {
+			@apply w-full h-40 p-2 border border-gray-300 rounded-md mb-4;
+		}
+	}
 }
